+++ src/test/CMakeLists.txt
@@ -113,6 +113,8 @@ set(
     tasks/operator_task_test.cpp
     testing_assert.cpp
     testing_assert.hpp
+    testing_utils.cpp
+    testing_utils.hpp
     utils/cuckoo_hashtable_test.cpp
     utils/numa_memory_resource_test.cpp
 )
