+++ src/test/operators/export_csv_test.cpp
@@ -5,6 +5,7 @@
 #include <utility>
 
 #include "../base_test.hpp"
+#include "../testing_utils.hpp"
 #include "gtest/gtest.h"
 
 #include "import_export/csv_meta.hpp"
@@ -15,7 +16,6 @@
 #include "storage/storage_manager.hpp"
 #include "storage/table.hpp"
 #include "utils/assert.hpp"
-#include "utils/unique_random_name.hpp"
 
 namespace opossum {
