+++ src/test/testing_utils.cpp
@@ -1,4 +1,4 @@
-#include "unique_random_name.hpp"
+#include "testing_utils.hpp"
 
 namespace opossum {
